+++ src/Annotations/Attributes/TestSuit.php
@@ -1,0 +1,20 @@
+<?php
+declare(strict_types=1);
+
+namespace MyTester\Annotations\Attributes;
+
+/**
+ * TestSuit attribute
+ *
+ * @author Jakub Konečný
+ * @internal
+ */
+#[Attribute(Attribute::TARGET_CLASS)]
+final class TestSuit {
+  public string $value;
+
+  public function __construct(string $value) {
+    $this->value = $value;
+  }
+}
+?>