+++ src/TestCase.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace MyTester;
 
 use MyTester\Annotations\NetteReflectionEngine;
+use MyTester\Annotations\PhpAttributesEngine;
 use MyTester\Annotations\Reader;
 use ReflectionClass;
 
@@ -34,6 +35,7 @@ abstract class TestCase {
 
   public function __construct() {
     $this->annotationsReader = new Reader();
+    $this->annotationsReader->registerEngine(new PhpAttributesEngine());
     $this->annotationsReader->registerEngine(new NetteReflectionEngine());
     $this->skipChecker = new SkipChecker($this->annotationsReader);
     $this->shouldFailChecker = new ShouldFailChecker($this->annotationsReader);