+++ tests/Annotations/NetteReflectionEngineTest.php
@@ -3,24 +3,21 @@ declare(strict_types=1);
 
 namespace MyTester\Annotations;
 
-use MyTester\Annotations\Attributes\Fail;
-use MyTester\Annotations\Attributes\TestSuit;
 use MyTester\ShouldFailChecker;
 use MyTester\TestCase;
 
 /**
- * Test suite for class Reader
+ * Test suite for class NetteReflectionEngine
  *
- * @testSuit Reader
+ * @testSuit NetteReflectionEngine
  * @author Jakub Konečný
  */
-#[TestSuit("Reader")]
-final class ReaderTest extends TestCase {
+final class NetteReflectionEngineTest extends TestCase {
   private function getAnnotationsReader(): Reader {
     static $annotationsReader = null;
     if($annotationsReader === null) {
       $annotationsReader = new Reader();
-      $annotationsReader->registerEngine(new DummyEngine());
+      $annotationsReader->registerEngine(new NetteReflectionEngine());
     }
     return $annotationsReader;
   }
@@ -34,15 +31,14 @@ final class ReaderTest extends TestCase {
 
   public function testGetAnnotation(): void {
     $this->assertNull((new Reader())->getAnnotation(TestCase::ANNOTATION_TEST_SUIT, static::class));
-    $this->assertSame("abc", $this->getAnnotationsReader()->getAnnotation(TestCase::ANNOTATION_TEST_SUIT, static::class));
+    $this->assertSame("NetteReflectionEngine", $this->getAnnotationsReader()->hasAnnotation(TestCase::ANNOTATION_TEST_SUIT, static::class));
     $this->assertNull((new Reader())->getAnnotation(ShouldFailChecker::ANNOTATION_NAME, static::class, "method"));
-    $this->assertSame("abc", $this->getAnnotationsReader()->getAnnotation(ShouldFailChecker::ANNOTATION_NAME, static::class, "method"));
+    $this->assertSame(1, $this->getAnnotationsReader()->getAnnotation(ShouldFailChecker::ANNOTATION_NAME, static::class, "method"));
   }
 
   /**
    * @fail(1)
    */
-  #[Fail(1)]
   private function method(): void {
   }
 }